added forgot password feature (read desc.)
users can now reset their password through a link sent to their email (requested via the forgot password modal on the login page). the backend generates a token stored in the user object (resetPasswordToken and resetPasswordExpires) with a 15 minute validity. there's also a validateToken function in the backend which checks if the token requested in the params exists at the users table at all, if not it throws a 400 error