added two-factor authentication (2FA) with OTP codes sent through email for enhanced security
added field 'twoFactorEnabled' in user model for this to work. when a user enables 2FA, they must verify a generated OTP code sent to their email. once 2FA is enabled, they will be prompted to enter an OTP code which is valid for 10 minutes, each time they log in. if the user decides to disable 2FA, they must also provide a valid OTP sent to their email to disable 2FA