added 4 new roles, added dark theme to the dashboard and added field description to role model
the 4 new roles are:
'customerSupport' which will receive the contact details that a user makes in his email so that he can reply to them
'orderManager' which can read and update each order made through the admin dashboard and will receive an email with the projected inventoryCount of the ordered items
'contentManager' which can create, read, update and delete slideshow images and FAQs through the admin dashboard
'productManager' which can create, read, update and delete products, categories, subcategories, subSubcategories, product restock subscriptions. suppliers and reviews
TODO: in the future i should be adding 'permissions' field as an array to the role model (e.g. for the role 'productManager' the permission field can be permissions: 'products', 'categories', 'subcategories' and so on) so that the permission don't get set in the App.jsx, but in the backend role table

also added dark theme to the dashboard to enhance UX. this is done through a dashboardTheme.js with the appropriate colors and ThemeContext.jsx to toggle between dark and light mode in the admin dashboard (state that gets saved in localStorage)